feat: new option to show Glow in same window
The option `in_place` determines if the output from Glow has to be
displayed in the same window as the input buffer or instead of
in a preview window like it was before this commit.

The ability to manage multiple windows with different markdown buffers
was incidentally implemented.

The `in_place_state` table is a map where:
- a key represent a certain window
- its value contains both the input buffer and the output buffer (the
  one created by *glow.nvim* holding the content of Glow).

```lua
in_place_state = {
    [WIN_1] = { BUFFER_BEFORE, BUFFER_CREATED },
    ...
}
```

So to the steps now are:
- Check if the config flag is enabled
- If it is save the state to retrieve BUFFER_BEFORE when needed
- Replace the buffer in WIN_1 with BUFFER_CREATED
- When "closing the window" reset the BUFFER_BEFORE in the WIN_1 and set
  in_place_state[WIN_1] to NIL

Line 346 had to be changed as well or:
- `Glow|Glow!|Glow` wouldn't have been possible (e.g. use command
  multiple times)
- `Glow|Glow` printed an error:

```lua
(current_win == win and not glow.config.in_place) -- keep previous
behaviour
or
in_place_state[current_win] -- Window is managed by Glow and needs to be
considered as a `glowpreview` buffer (e.g. closed or not bothered)
```

This is because before `current_win == win` always meant we were on the
preview window which now isn't the case anymore.